backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
ultralytics==8.3.0
celery==5.4.0
redis==5.0.7
gunicorn==22.0.0
django-celery-results==2.5.1
